remote respect-parenthesis and append NULL
we don't need the --respect-parenthesis param

Aand the unmatched \) is actually the wrong solution to this error
message:

	convert: unbalanced parenthesis `)' ...

... which actually happens because a command must end with an image
name, without a preceeding write.

Instead, adding NULL: to the end of eh command works fine.

The lack of a NULL: also seems to cause the spurious files like ")-0"
and ")-1" to be written. This in turn makes the convert take twice as
long as it needs to.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>