Use gthumb if fbv not installed
If the user is on a X based system, they can run photobooth without
using fbv and rely on gthumb to display the images.

This also relies on the user having gthumb installed, but also setting
it up to enable Preferences->"Reuse the active window to open files".

Because we then rely on gthumb to be running in full screen mode and in
foreground, we kill all existing instances of gthumb and respawn it.

However, on a headless system with /dev/fb0, the user can install fbv to
display the final files as per previous behaviour.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>