cycle through backgrounds
This will cycle through background images in the resources folder that
match the naming pattern:

  background-*.jpg

The backgrounds will be used in alphanumberic order as shown by a
default "ls".

The state file is stored in the resources directory to show which image
to use next. This is needed because showjpg is called from fresh every
time we display a picture and it cannot remember state internally.

When we get to the end of the list of backgrounds, it cycles back to the
start again.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>